drivers/docker: move lifted code out to separate file and link the source & license
